requirements.txt
flask
gunicorn
flask-cors
requests
python-dotenv
fuzzywuzzy
python-Levenshtein